% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_layers.R
\name{combine_layers}
\alias{combine_layers}
\title{Combine layers}
\usage{
combine_layers(graph1, graph2 = NULL, interaction.df = NULL)
}
\value{
a graph
}
\description{
Return merged graph from two graph layers
}
\examples{
data(BIOGRID)
db <- BIOGRID
X <- V(db)$name[1:10]
biogrid.res <- get_interaction_from_database(X, db, type = "db", user.ego = FALSE)
}
