% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions.R
\name{get_go_info}
\alias{get_go_info}
\title{Get GO info}
\usage{
get_go_info(go)
}
\arguments{
\item{go}{a character, GO term}
}
\value{
a data.frame with the following columns: "GOID", "DEFINITION", "ONTOLOGY", "TERM"
}
\description{
From a GO terms (GOID), return definition, ontology and term values from GO.db
}
