% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grn.R
\name{get_grn}
\alias{get_grn}
\title{Get Gene Regulatory Network from dataset}
\usage{
get_grn(X, cluster = NULL, method = c("aracne"))
}
\arguments{
\item{method}{}
}
\description{
A more detailed description
}
\examples{
data(HeLa)
# grn only on gene
cluster.mRNA <- timeOmics::getCluster(HeLa$getCluster, user.block = "mRNA")
X <- HeLa$raw$mRNA
grn.res <- get_grn(X = HeLa$raw$mRNA, cluster = cluster.mRNA, method = "aracne")

}
