% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions.R
\name{get_interaction_from_ORA}
\alias{get_interaction_from_ORA}
\title{Get interaction from ORA enrichment analysis}
\usage{
get_interaction_from_ORA(
  query,
  sources = "GO",
  organism = "hsapiens",
  signif.value = TRUE
)
}
\arguments{
\item{query}{a vector (or a list) of character with the ID to perform
the ORA analysis}

\item{sources}{(optional) a character in
(GO, KEGG, REAC, TF, MIRNA, CORUM, HP, HPA, WP)}

\item{organism}{(optional) a character (default = 'hsapiens')}

\item{signif.value}{(optional) a logical, default = ''}
}
\value{
a graph object (or list of graph) containing the interaction between
the query and the target terms.
}
\description{
Returns results of an ORA analysis as an interaction graph
}
\examples{
query <- c('IL15', 'CDHR5', 'TGFA', 'C4B')
get_interaction_from_ORA(query,
                         sources = 'GO')

query <- list('All' = c('IL15', 'CDHR5', 'TGFA', 'C4B'),
              'c1' = c('IL15', 'CDHR5', 'TGFA'))
get_interaction_from_ORA(query,
                         sources = 'GO')

}
\seealso{
\code{\link[gprofiler2]{gost}}  \code{\link[gprofiler2]{gconvert}}
}
