% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions.R
\name{get_ORA}
\alias{get_ORA}
\title{ORA enrichment analysis}
\usage{
get_ORA(query, sources = NULL, organism = "hsapiens")
}
\arguments{
\item{query}{a vector of character, a lit of ID}

\item{sources}{a character or list of character}

\item{organism}{a character (default = 'hsapiens')}
}
\value{
a data.frame containing the enrichment result
}
\description{
Returns results of an ORA analysis
}
\seealso{
\code{\link[gprofiler2]{gost}}
}
