% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_add_cluster.R
\name{graph_add_cluster}
\alias{graph_add_cluster}
\title{Add cluster label to graph nodes}
\usage{
graph_add_cluster(X, cluster)
}
\description{
plpl
}
\examples{
cluster <- timeOmics::getCluster(HeLa$getCluster, user.block = "mRNA")
X <- HeLa$raw$mRNA

}
