% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeLa.R
\docType{data}
\name{HeLa}
\alias{HeLa}
\title{HeLa Cell Cycling dataset}
\format{
a list of data.frame
\describe{
\item{raw}{data.frame, raw data}
\item{modelled}{data.frame, modelled data}
}
}
\usage{
HeLa
}
\description{
A list of data.frames
}
\keyword{datasets}
