% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmp_T2D.R
\docType{data}
\name{hmp_T2D}
\alias{hmp_T2D}
\title{HeLa Cell Cycling dataset}
\format{
a list of data.frame
\describe{
\item{raw}{data.frame, raw data}
\item{modelled}{data.frame, modelled data}
}
}
\usage{
hmp_T2D
}
\description{
A list of data.frames
}
\keyword{datasets}
