% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_layers_by_correlation.R
\name{interaction_from_correlation}
\alias{interaction_from_correlation}
\title{Interaction_from_correlation}
\usage{
interaction_from_correlation(X, Y, threshold = 0.5)
}
\arguments{
\item{X}{a data.frame or list of data.frame (with a similar number of row).}

\item{Y}{a data.frame or list of data.frame (with a similar number of row).}

\item{threshold}{a threshold to cut the correlation matrix above which a link is created between a feature from X and a feature from Y.}
}
\value{
an 'igraph' object
}
\description{
Compute correlation between two dataframe X and Y (or list of data.frame).
An incidence graph is returned. A link between two features is produced if their correlation (absolute value) is above the threshold.
}
\examples{
X <- matrix(rexp(200, rate=.1), ncol=20)
Y <- matrix(rexp(200, rate=.1), ncol=20)
interaction_from_correlation(X,Y)

X <- list(matrix(rexp(200, rate=.1), ncol=20), matrix(rexp(200, rate=.1), ncol=20))
Y <- matrix(rexp(200, rate=.1), ncol=20)
interaction_from_correlation(X,Y)
}
