% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWR.R
\name{rwr_find_closest_type}
\alias{rwr_find_closest_type}
\title{RWR Find closest nodes}
\usage{
rwr_find_closest_type(X, seed = NULL, attribute = NULL, value = NULL)
}
\arguments{
\item{X}{a random walk result from \code{random_walk_restart}}

\item{seed}{a character vector or NULL. If NULL, all the seeds from X are considered.}

\item{attribute}{a character value or NULL. If NULL, the closest node is returned.}

\item{value}{a character value or NULL. If NULL, the closest node for a given attribute is returned.}
}
\value{
A list of \code{data.frame} for each seed containing the closest nodes per seed and their vertex attributes.
If X is \code{list.rwr}, the returned value is a list of list.
}
\description{
This function returns the closest nodes from a seed with a given attribute and value.
In the random walk context, it might be useful to get the closest Gene Ontology annotation nodes from unannotated seeds.
}
\examples{
data(HeLa)
X <- HeLa$raw$mRNA
cluster.mRNA <- timeOmics::getCluster(HeLa$getCluster, user.block = "mRNA")
grn.res <- get_grn(X = HeLa$raw$mRNA, cluster = cluster.mRNA, method = "aracne")

X <- grn.res[["All"]]
all_seed <- igraph::V(X)$name
seed <- all_seed[1:5]

rwr_res <- random_walk_restart(X, seed)

rwr_res <- random_walk_restart(grn.res, seed)
rwr_closest_res <- rwr_find_closest_type(rwr_res, attribute = "cluster")
rwr_closest_res <- rwr_find_closest_type(rwr_res)

}
