% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_graph_stats.R
\name{get_graph_stats}
\alias{get_graph_stats}
\title{get graph statistics}
\usage{
get_graph_stats(X)
}
\description{
A more detailed description
}
\examples{
data(HeLa)
# grn only on gene
cluster.mRNA <- timeOmics::getCluster(HeLa$getCluster, user.block = "mRNA")
X <- HeLa$raw$mRNA
grn.res <- get_grn(X = HeLa$raw$mRNA, cluster = cluster.mRNA, method = "aracne")
get_graph_stats(grn.res)


}
