% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_interaction_from_database.R
\name{get_interaction_from_database}
\alias{get_interaction_from_database}
\title{Get interaction from database}
\usage{
get_interaction_from_database(X, db = NULL, type = "db", user.ego = FALSE)
}
\arguments{
\item{X}{vector of nodes or list of vectors}

\item{db}{data.frame (with two columns: from, to) or igraph}

\item{type}{character added to node metadata}

\item{user.ego}{logical, if user.ego == TRUE looks for first degree neighbors in db and add 'mode' metadata ('core'/'extended')}
}
\value{
a subset graph of db from X list of nodes
}
\description{
Returns an interaction graph from a vector of nodes (or a list of vectors) and an interaction database (data.frame or igraph)
}
\examples{
data(BIOGRID)
db <- BIOGRID
X <- V(db)$name[1:10]
biogrid.res <- get_interaction_from_database(X, db, type = "db", user.ego = FALSE)

X <- letters[1:5]
db <- as.data.frame(list(from = c(letters[1:5], letters[11:15]), to = c(letters[1:10])))
}
