% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmp_T2D.R
\docType{data}
\name{hmp_T2D}
\alias{hmp_T2D}
\title{hmp_T2D}
\format{
a list of data.frame
\describe{
\item{raw}{data.frame, raw data}
\item{modelled}{data.frame, modelled data}
\item{getCluster.res}{data.frame, clustering results from timeOmics}
\item{getCluster.sparse.res}{data.frame, sparse clustering results from timeOmics}
\item{interaction.biogrid}{data.frame, interactions from BioGRID database}
\item{interaction.TF}{data.frame, TFome interactions from TTrust and TF2DNA}
\item{medlineranker.res.df}{data.frame, medlineRanker enrichment results}
\item{graph.gut}{list of igraph, gut graph obtained with SparCC}
}
}
\usage{
hmp_T2D
}
\description{
A list of data.frames
}
\keyword{datasets}
