% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWR.R
\name{random_walk_restart}
\alias{random_walk_restart}
\title{Random Walk with Restart}
\usage{
random_walk_restart(X, seed, r = 0.7)
}
\arguments{
\item{X}{an igraph or list.igraph object.}

\item{seed}{a character vector. Only seeds present in X are considered.}

\item{r}{a numeric value between 0 and 1. It sets the probability of restarting to a seed node after each step.}
}
\value{
Each element of X returns a list (class = 'rwr') containing the following elements:
\item{rwr}{a \code{data.frame}, the RWR results for each valid seed.}
\item{seed}{a character vector with the valid seeds}
\item{graph}{\code{igraph} object from X}
If X is a \code{list.igraph}, the returned object is a \code{list.rwr}.
}
\description{
This function performs a propagation analysis by random walk with restart in a multi-layered network from specific seeds.
}
\examples{

data(HeLa)
X <- HeLa$raw$mRNA
cluster.mRNA <- timeOmics::getCluster(HeLa$getCluster, user.block = "mRNA")
grn.res <- get_grn(X = HeLa$raw$mRNA, cluster = cluster.mRNA, method = "aracne")

X <- grn.res
Xi <- grn.res[["Up_Up"]]
seed <- igraph::V(Xi)$name[1:5]
rwr_res <- random_walk_restart(Xi, seed)

rwr_res <- random_walk_restart(X, seed)

random_walk_restart(X, seed = "A")

}
\seealso{
\code{\link[RandomWalkRestartMH]{Random.Walk.Restart.Multiplex}}, \code{\link[netOmics]{rwr_find_seeds_between_attributes}}, \code{\link[netOmics]{rwr_find_closest_type}}
}
