% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWR.R
\name{rwr_find_seeds_between_attributes}
\alias{rwr_find_seeds_between_attributes}
\title{RWR Find seeds between attributes}
\usage{
rwr_find_seeds_between_attributes(X, seed = NULL, k = 15, attribute = "type")
}
\arguments{
\item{X}{a random walk result from \code{random_walk_restart}}

\item{seed}{a character vector or NULL. If NULL, all the seeds from X are considered.}

\item{attribute}{a character value or NULL. If NULL, the closest node is returned.}
}
\value{
A list of igraph object for each seed.
If X is a list, it returns a list of list of graph.
}
\description{
This function returns a subgraph if any vertex shares different attributes value.
It can be useful to identify vertex shared between clusters or omics types.
}
\examples{
data(HeLa)
X <- HeLa$raw$mRNA
cluster.mRNA <- timeOmics::getCluster(HeLa$getCluster, user.block = "mRNA")
grn.res <- get_grn(X = HeLa$raw$mRNA, cluster = cluster.mRNA, method = "aracne")

X <- grn.res[["All"]]
all_seed <- igraph::V(X)$name
seed <- c("AAK1","ABCC1","ABL2")

rwr_res <- random_walk_restart(X, seed)
rwr_cluster_res <- rwr_find_seeds_between_attributes(rwr_res, attribute = "cluster")
rwr_cluster_res <- rwr_find_seeds_between_attributes(rwr_res, attribute = "block")

rwr_res <- random_walk_restart(grn.res, seed)
rwr_cluster_res <- rwr_find_seeds_between_attributes(rwr_res)

rwr_res <- random_walk_restart(X, all_seed)
rwr_cluster_res <- rwr_find_seeds_between_attributes(X = rwr_res, attribute = "cluster")
rwr_block_res <- rwr_find_seeds_between_attributes(X = rwr_res, attribute = "block")#' 

}
