% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWR.R
\name{rwr_find_seeds_between_attributes}
\alias{rwr_find_seeds_between_attributes}
\title{RWR Find seeds between clusters}
\usage{
rwr_find_seeds_between_attributes(
  X,
  seed = NULL,
  k = 15,
  attribute = "cluster"
)
}
\arguments{
\item{X}{a random walk result from \code{random_walk_restart}}

\item{seed}{a character vector or NULL. If NULL, all the seeds from X are considered.}

\item{attribute}{a character value or NULL. If NULL, the closest node is returned.}
}
\value{
A list of \code{data.frame} for each seed containing the closest nodes per seed and their vertex attributes.
If X is \code{list.rwr}, the returned value is a list of list.
}
\description{
RWR Find seeds between clusters
}
\examples{
#' data(HeLa)
X <- HeLa$raw$mRNA
cluster.mRNA <- timeOmics::getCluster(HeLa$getCluster, user.block = "mRNA")
grn.res <- get_grn(X = HeLa$raw$mRNA, cluster = cluster.mRNA, method = "aracne")

X <- grn.res[["All"]]
all_seed <- igraph::V(X)$name
seed <- all_seed[1:5]

rwr_res <- random_walk_restart(X, seed)
rwr_cluster_res <- rwr_find_seeds_between_attributes(rwr_res, attribute = "cluster")
rwr_cluster_res <- rwr_find_seeds_between_attributes(rwr_res, attribute = "block")

rwr_res <- random_walk_restart(grn.res, seed)
rwr_cluster_res <- rwr_find_seeds_between_attributes(rwr_res)

rwr_res <- random_walk_restart(X, all_seed)
rwr_cluster_res <- rwr_find_seeds_between_attributes(X = rwr_res, attribute = "cluster")
rwr_res <- random_walk_restart(X, all_seed, k = 15)
rwr_cluster_res <- rwr_find_seeds_between_attributes(rwr_res)

}
